gconvert: Correctly annotate string types and output parameters
Note that the iconv wrapper API works with byte arrays. It's wrong to
default to utf8 there, because iconv can read and produce strings with
interior nul characters which are not allowed in (type utf8).
The documentation was misleading about that in some places, so that got
corrected as well.

https://bugzilla.gnome.org/show_bug.cgi?id=756128