gcharset, gconvert: More robust UTF-8 detection
Whether the charset is UTF-8 was decided by searching "UTF-8"
as a substring in the platform's best trim of a string that is supposed
to represent the charset name.
This is problematic for two reasons:
1. The search is case-sensitive, so this does not work on Mac OS X,
where nl_langinfo(CODESET) returns "utf-8".
2. The unanchored substring match is a bit too loose. Even though it
likely works in all cases where it should, and all applications should
do UTF-8 validation also in case of a positive match, there is more
space for potential security exploits than is necessary for a practical
optimization check.

The check is changed to a case-insensitive suffix match. That should
cover any old platforms without nl_langinfo() where either "CHARSET"
or "language_COUNTRY.CHARSET" locale specification format is used to
select the charset. A @blah modifier suffix is ignored if found.

https://bugzilla.gnome.org/show_bug.cgi?id=738044